Move EditEvent to separate EventForm component
